+++ commands/update.go
@@ -1,7 +1,9 @@
 package commands
 
 import (
+	"bufio"
 	"fmt"
+	"strings"
 
 	"github.com/spf13/cobra"
 )
@@ -16,6 +18,24 @@ func newUpdateCommand(ctx *Context) *cobra.Command {
 		Use:   "update",
 		Short: "updates an existing password",
 		RunE: func(cmd *cobra.Command, args []string) error {
+			reader := bufio.NewReader(cmd.InOrStdin())
+			if len(machine) == 0 {
+				fmt.Fprintf(cmd.OutOrStdout(), "Machine: ")
+				line, err := reader.ReadString('\n')
+				if err != nil {
+					return fmt.Errorf("ReadString() failed: %s", err)
+				}
+				machine = strings.TrimSuffix(line, "\n")
+			}
+			if len(user) == 0 {
+				fmt.Fprintf(cmd.OutOrStdout(), "User: ")
+				line, err := reader.ReadString('\n')
+				if err != nil {
+					return fmt.Errorf("ReadString() failed: %s", err)
+				}
+				user = strings.TrimSuffix(line, "\n")
+			}
+
 			generatedPassword := password
 			if len(password) == 0 {
 				var err error
@@ -48,11 +68,9 @@ func newUpdateCommand(ctx *Context) *cobra.Command {
 			return nil
 		},
 	}
-	cmd.Flags().StringVarP(&machine, "machine", "m", "", "machine (required)")
-	cmd.MarkFlagRequired("machine")
+	cmd.Flags().StringVarP(&machine, "machine", "m", "", "machine (default: ask)")
 	cmd.Flags().StringVarP(&service, "service", "s", "http", "service")
-	cmd.Flags().StringVarP(&user, "user", "u", "", "user (required)")
-	cmd.MarkFlagRequired("user")
+	cmd.Flags().StringVarP(&user, "user", "u", "", "user (default: ask)")
 	cmd.Flags().StringVarP(&password, "password", "p", "", "new password")
 	cmd.Flags().VarP(&passwordType, "type", "t", `password type ("plain" or "totp")`)