+++ guide/src/news.md
@@ -4,6 +4,7 @@
 
 - create: add interactive mode in case `--machine` or `--user` is not specified
 - search: add interactive mode when no search terms are specified
+- update: add interactive mode in case `--machine` or `--user` is not specified
 
 ## 5.0