+++ guide/src/news.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## main
+## 6.0
 
 - create, update and delete: add interactive mode in case `--machine` or `--user` is not specified
 - search: add interactive mode when no search terms are specified